Code clean up/Moved password hashing to models.py
When creating a user from sample data I did not have an easy way
to insert a hashed password into the database. By having the model hash
anything passed into the constructor I was able to mitigate this
problem. I also added the option to specify the id of a user when it is
being created (only a developer tool but needs testing)