Implements seeding using only the database connection
Moves away from seeding using the REST api. Could not find an elegant
way to handle all cases. There does not seem to be a way to call the
constructors of all the database's models when bulk_inserting into the
database. This code should be refactored in the future should such a
method be found.